first draft of moving to ed25519
This commit removes the password hash approach in favor of ED25519
signature verification scheme.

The contract now assumes the `pubkey` stored in the datum is achieved
using the following seed:
`[ nonce, username, raw_password ]`

Or perhaps even something like:
`[ nonce, [ username ], [ raw_password ] ]`

Where `[ a ]` is used to notate hash of `a`. The on-chain code does not
enforce any scheme for the underlying seed.

It also doesn't need to enforce the pubkey to be changed for each
withdrawal, and offers a separate endpoint for doing so.