first draft of support for contributors
The new contract, `treasury`, is meant to serve as a funding solution
supported by contributors.

It allows withdrawal of a single UTxO for each new account. The
off-chain code is meant to build contribution transactions in such a way
that each produced UTxO at treasury will hold enough funds for the 2
UTxOs of an account.

The datums of these UTxOs also hold the addresses of the contributors,
so that the contracts can validate the staking parts of the UTxOs to be
that of the contributors.

`treasury` is still incomplete.